cleanup and some docs
